Fixes compilation errors on Windows and adds SPIRV-Cross subdirectory also for OpenGl builds

+++ include/OpenGl/erm/rendering/renderer/RenderContext.h
@@ -16,7 +16,7 @@ namespace erm {
 	public:
 		RenderContext();
 
-		void Draw(DrawMode drawMode, unsigned int count) const;
+		void Draw(DrawMode drawMode, uint64_t count) const;
 
 		void Clear() const;
